+++ lib/fast_ostruct.rb
