+++ spec/lib/fast_ostruct_spec.rb
