+++ spec/spec_helper.rb
@@ -3,7 +3,7 @@
 require 'bundler/setup'
 require 'pry'
 require 'rack/test'
-require 'fast_open_struct'
+require 'fast_ostruct'
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
